8mo.26: FourCC-wobble drops root-caused — two were real defects, now fixed (+2 exact)
The three 100%-to-partial drops the FourCC catalog commit (984933ab1) was
blamed for decompose as:

- 0x52c990 map_overlay_geometry: recovered to 100% on its own during this
  session's rebuilds — pure link-order wobble, as originally suspected.
- 0x4e4630 TMinor::SumDiplomacyState1c6AndRelationDeltaSnapshot: a REAL
  pre-existing defect the wobble label hid. The whole virtual (TCountry slot
  0x1c family) returns SHORT, not int — the originals load 16-bit halves with
  no sign extension and only AX is meaningful (TGreatPower 0x4dda60's
  mov ax/mov cx/add eax,ecx shape). Signature corrected across TCountry/
  TMinor/TGreatPower; TMinor and TCountry bodies now EXACT.
- 0x49ace0 ResolvePaletteIndexColor: single commutative OR-chain load-order
  line ([+5] green vs [+6] blue first) — the documented phantom-drop class,
  benign; likewise TGreatPower's own Sum body (60%, add-chain operand order,
  natural expression kept per the commutative_order guidance).

Verdict recorded per the bead's acceptance: the catalog restructuring itself
is benign wobble, but 'wobble' had absorbed two genuine defects; net +2 exact.
Also refreshed the stale scalar-type-audit and turn-event-coverage reports the
gates flagged. Gates green, baselines updated.

+++ include/game/nation/TGreatPower.h
@@ -61,7 +61,7 @@ public:
   void AddRegionIdToNationOwnedRegionList(int regionId) override;
   void SetNationPercentFieldByModeAndDescriptorLinks(int targetNationSlot, int policyCode) override;
   void DecrementDiplomacyCounterA2ByValue(int delta) override;
-  int SumDiplomacyState1c6AndRelationDeltaSnapshot(short nationSlot) override; // slot 0x1c
+  short SumDiplomacyState1c6AndRelationDeltaSnapshot(short nationSlot) override; // slot 0x1c
   short GetDiplomacyCounterA2(void) override;                                  // slot 0x1d
   short GetDiplomacyExternalStateByTarget(short nationSlot) override;          // slot 0x1e
   short QueryNationMetricBySlot7C(short metricSlot) override;                  // slot 0x1f
